feat: release 0.3.5 with daily word notifications
- Introduced opt-in daily word notifications with a customizable reminder time via Settings.
- Implemented WorkManager to schedule notifications, ensuring reminders persist after app launches and device reboots.
- Added a debug test notification feature for easier testing in debug builds.
- Updated app version to **0.3.5** (versionCode **12**) in build configuration and documentation.

+++ app/src/main/java/com/example/wordofday/data/repository/WordRepository.kt
@@ -185,7 +185,9 @@ internal fun selectionIndex(
     salt: Int = 0,
 ): Int {
     if (poolSize <= 0) return 0
-    var hash = dayOfYear xor (rotationOffset * 0x9E3779B9) xor (salt * 0x85EBCA6B)
+    var hash = dayOfYear xor
+        (rotationOffset * 0x9E3779B9.toInt()) xor
+        (salt * 0x85EBCA6B.toInt())
     hash = hash xor (hash shl 16)
     hash = hash xor (hash ushr 16)
     hash *= 0x7FEB352D
